Fix loading of already-inline images
If the src is already a 'data:' URL, we currently only check if it's not
the loading image; if so, we assume the image has already loaded.  But
some pages have img.src already as a 'data:' URL, as well as srcset
attribute. With the current scheme above, we'll never load the image.

With this change we also check the img.originalUrl, to make sure that we
did attempt to load the URL. Thus we're able to load the actual image
(using the srcset attribute).